refactor: move component transformers to one folder
