+++ __tests__/transformers/mdxish-inline-component-blocks.test.ts
@@ -8,8 +8,8 @@ import { describe, expect, it } from 'vitest';
 
 import { mdxComponentFromMarkdown } from '../../lib/mdast-util/mdx-component';
 import { mdxComponent } from '../../lib/micromark/mdx-component';
-import mdxishComponentBlocks from '../../processor/transform/mdxish/mdxish-component-blocks';
-import mdxishInlineComponentBlocks from '../../processor/transform/mdxish/mdxish-inline-component-blocks';
+import mdxishComponentBlocks from '../../processor/transform/mdxish/components/component-blocks';
+import mdxishInlineComponentBlocks from '../../processor/transform/mdxish/components/inline-component-blocks';
 import { collectNodes } from '../helpers';
 
 // Mirrors the relevant slice of the mdxish pipeline: tokenizer + both
