+++ lib/mdxish.ts
@@ -26,17 +26,17 @@ import codeTabsTransformer from '../processor/transform/code-tabs';
 import embedTransformer from '../processor/transform/embeds';
 import imageTransformer from '../processor/transform/images';
 import { closeSelfClosingHtmlTags } from '../processor/transform/mdxish/close-self-closing-html-tags';
+import mdxishComponentBlocks from '../processor/transform/mdxish/components/component-blocks';
+import mdxishInlineComponentBlocks from '../processor/transform/mdxish/components/inline-component-blocks';
+import mdxishInlineComponents from '../processor/transform/mdxish/components/inline-components';
+import mdxishSelfClosingBlocks from '../processor/transform/mdxish/components/self-closing-blocks';
+import { processSnakeCaseComponent } from '../processor/transform/mdxish/components/snake-case-components';
 import evaluateExpressions from '../processor/transform/mdxish/evaluate-expressions';
 import generateSlugForHeadings from '../processor/transform/mdxish/heading-slugs';
 import magicBlockTransformer from '../processor/transform/mdxish/magic-blocks/magic-block-transformer';
-import mdxishComponentBlocks from '../processor/transform/mdxish/mdxish-component-blocks';
 import mdxishHtmlBlocks from '../processor/transform/mdxish/mdxish-html-blocks';
-import mdxishInlineComponentBlocks from '../processor/transform/mdxish/mdxish-inline-component-blocks';
-import mdxishInlineComponents from '../processor/transform/mdxish/mdxish-inline-components';
 import mdxishJsxToMdast from '../processor/transform/mdxish/mdxish-jsx-to-mdast';
 import mdxishMermaidTransformer from '../processor/transform/mdxish/mdxish-mermaid';
-import mdxishSelfClosingBlocks from '../processor/transform/mdxish/mdxish-self-closing-blocks';
-import { processSnakeCaseComponent } from '../processor/transform/mdxish/mdxish-snake-case-components';
 import mdxishTables from '../processor/transform/mdxish/mdxish-tables';
 import mdxishTablesToJsx from '../processor/transform/mdxish/mdxish-tables-to-jsx';
 import { normalizeCompactHeadings } from '../processor/transform/mdxish/normalize-compact-headings';
